about the application
/***************************about the application*************************************
The app will allow the user to enter a number and view its decomposition.           *
The app will contain the following sections:App titles and text field.              *
A text field (in which a placeholder will appear with the string "positive integer")*
 to enter the number. Check box for creating a decomposition tree.                  *
Button calculate Clear button The projection area of ​​the answer.                    *
*************************************************************************************/